==> notebooks/03_feature_engineering_client.ipynb <==
# 03 - Feature Engineering Client


# 03 — Feature Engineering Client (Olist)

Objectif :
Construire une table client agrégée (1 ligne = 1 client unique)
à partir des données transactionnelles nettoyées,
afin de permettre la segmentation et la modélisation prédictive.


# Chargement des tables nettoyés

import pandas as pd
from pathlib import Path

pd.set_option("display.max_columns", 200)
pd.set_option("display.width", 120)

PROJECT_DIR = Path.cwd().parent
INTERIM_DIR = PROJECT_DIR / "data" / "interim"


customers = pd.read_parquet(INTERIM_DIR / "customers_clean.parquet")
orders = pd.read_parquet(INTERIM_DIR / "orders_delivered.parquet")
order_items = pd.read_parquet(INTERIM_DIR / "order_items_clean.parquet")
payments = pd.read_parquet(INTERIM_DIR / "payments_clean.parquet")
reviews = pd.read_parquet(INTERIM_DIR / "reviews_clean.parquet")
products = pd.read_parquet(INTERIM_DIR / "products_clean.parquet")


# Vérification rapide

print("Customers :", customers.shape)
print("Orders    :", orders.shape)
print("Items     :", order_items.shape)
print("Payments  :", payments.shape)
print("Reviews   :", reviews.shape)
print("Products  :", products.shape)


# Creation de la table pivot

clients = customers[["customer_unique_id"]].drop_duplicates().copy()

clients.shape


# Jointures clients <> commandes

orders_clients = orders.merge(
    customers[["customer_id", "customer_unique_id"]],
    on="customer_id",
    how="left"
)


print(orders_clients[["order_id", "customer_unique_id"]].head())


Les jointures ont été réalisées en LEFT JOIN afin de préserver
l’exhaustivité des transactions e-commerce,
garantissant qu’aucune commande livrée ne soit exclue
lors de la construction des variables client.


# Calcule du RFM

# On repart de la table orders_clients créée précédemment
orders_clients = orders.merge(
    customers[["customer_id", "customer_unique_id"]],
    on="customer_id",
    how="left"
)


# Définir la date de référence (CRUCIAL)

reference_date = orders_clients["order_purchase_timestamp"].max()

reference_date


# Calcul de la RECENCY (R)

recency = (
    orders_clients
    .groupby("customer_unique_id")["order_purchase_timestamp"]
    .max()
    .reset_index()
)

recency["recency"] = (reference_date - recency["order_purchase_timestamp"]).dt.days

recency = recency[["customer_unique_id", "recency"]]


# Calcul de la FREQUENCY (F)

frequency = (
    orders_clients
    .groupby("customer_unique_id")["order_id"]
    .nunique()
    .reset_index(name="frequency")
)


# Calcul du MONETARY (M)

payments_per_order = (
    payments
    .groupby("order_id")["payment_value"]
    .sum()
    .reset_index()
)


# Jointure paiements <> Clients

payments_clients = payments_per_order.merge(
    orders_clients[["order_id", "customer_unique_id"]],
    on="order_id",
    how="left"
)


# Calcul du Monetary

monetary = (
    payments_clients
    .groupby("customer_unique_id")["payment_value"]
    .sum()
    .reset_index(name="monetary")
)


# Fusion des 3 composantes RFM

rfm = recency.merge(frequency, on="customer_unique_id", how="left")
rfm = rfm.merge(monetary, on="customer_unique_id", how="left")

rfm.head()


# Vérifications essentielles (OBLIGATOIRES)

# Dimensions
rfm.shape


# Valeurs manquantes
print(rfm.isna().mean())


# Statistiques descriptives
print(rfm.describe())


# Features supllémentaires

### Panier moyen par client

- Le client achète-t-il peu mais cher ou ou souvent mais peu cher ?

rfm["avg_basket"] = rfm["monetary"] / rfm["frequency"]


rfm[["monetary", "frequency", "avg_basket"]].head()


### Nombre total d'articles achetés

- le client achète-t-il plusieurs produits par commandes ?

# Calcule intermédiare items <> clients
items_clients = order_items.merge(
    orders_clients[["order_id", "customer_unique_id"]],
    on="order_id",
    how="left"
)


# Agrégation au niveau client
total_items = (
    items_clients
    .groupby("customer_unique_id")["order_item_id"]
    .count()
    .reset_index(name="total_items")
)


### Délai moyen de livraison 

- Le client subit-il des délais longs ?

# Calcule du délai par commande
orders_clients["delivery_delay_days"] = (
    orders_clients["order_delivered_customer_date"]
    - orders_clients["order_purchase_timestamp"]
).dt.days


# Agrégation au niveau client (varaible très liée à la satisfaction du client)
avg_delivery_delay = (
    orders_clients
    .groupby("customer_unique_id")["delivery_delay_days"]
    .mean()
    .reset_index(name="avg_delivery_delay")
)


### Score moyen de satisfaction client 

- Le client est-il globalement satisfait ?

# Jointure reviews <> clients
reviews_clients = reviews.merge(
    orders_clients[["order_id", "customer_unique_id"]],
    on="order_id",
    how="left"
)


# Agrégation au niveau client
# certains clients n'ont pas d'avis donc les valeurs manquantes sont normales
avg_review_score = (
    reviews_clients
    .groupby("customer_unique_id")["review_score"]
    .mean()
    .reset_index(name="avg_review_score")
)


### Diversité des catégories achetées

- Le client achète-t-il dans plusieurs univers produits ?

# Items <> produits
items_products = order_items.merge(
    products[["product_id", "product_category_name"]],
    on="product_id",
    how="left"
)


# items <> clients
items_products_clients = items_products.merge(
    orders_clients[["order_id", "customer_unique_id"]],
    on="order_id",
    how="left"
)


# Agrégations
category_diversity = (
    items_products_clients
    .groupby("customer_unique_id")["product_category_name"]
    .nunique()
    .reset_index(name="nb_categories")
)


### FUSION DE TOUS LES FEATURES CLEINTS

client_features = (
    rfm
    .merge(total_items, on="customer_unique_id", how="left")
    .merge(avg_delivery_delay, on="customer_unique_id", how="left")
    .merge(avg_review_score, on="customer_unique_id", how="left")
    .merge(category_diversity, on="customer_unique_id", how="left")
)


### Vérifications finales

# Dimensions
client_features.shape


# Valeurs manquantes 
print(client_features.isna().mean().sort_values(ascending=False))


### Aperçu

print(client_features.head())


## Feature Engineering — Conclusion

La table client finale a été construite au niveau du client unique.
Elle combine :
- des indicateurs RFM
- des variables comportementales
- des indicateurs logistiques
- une mesure de la satisfaction client

Cette table constitue la base de la segmentation et de la modélisation prédictive.


### Savegarde

# Sauvegarde de la table client finale
client_features.to_parquet(
    INTERIM_DIR / "client_features.parquet",
    index=False
)

print(" client_features sauvegardé dans data/interim/")
